Restructured JS modules
