+++ src/assets/styles/styles.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'logoFont';
-  src: url('./PermanentMarker-Regular.ttf') format('ttf');
+  src: url('../fonts/PermanentMarker-Regular.ttf') format('ttf');
 }
 
 body {
