--- src/generalTemplate.js
@@ -1,39 +1,0 @@
-// const body = document.querySelector('body');
-// const content = document.querySelector('#content');
-// const brandLogo = document.createElement('h1');
-
-// const loadHeader = () => {
-//   const header = document.createElement('HEADER');
-//   header.setAttribute('id', 'header');
-
-//   const workingHeader = document.createElement('DIV');
-//   workingHeader.setAttribute('id', 'workingHeader');
-
-//   workingHeader.classList.add('container', 'py-2');
-
-//   brandLogo.setAttribute('id', 'logo');
-//   brandLogo.innerText = 'TickyTick';
-
-//   workingHeader.append(brandLogo);
-//   header.append(workingHeader)
-
-//   const mainBody = document.createElement('DIV');
-//   mainBody.classList.add('container', 'row', 'mainBody', 'mx-auto');
-//   const filterSection = document.createElement('DIV')
-//   filterSection.classList.add('col-3', 'filterSection')
-
-//   const listSection = document.createElement('DIV')
-//   listSection.classList.add('col-6', 'listSection')
-
-//   const extraSection =document.createElement('DIV')
-//   extraSection.classList.add('col-3', 'extraSection')
-
-//   mainBody.append(filterSection, listSection, extraSection)
-
-//   content.append(mainBody);
-
-//   body.insertBefore(header, content);
-
-// };
-
-// export default loadHeader;
